Refactor FacebookCampaign class: - Add FbIgBase class, InstagramCampaign class - Add bulk update tests for Instagram
